Removed FromMemoryAlpha trait
Closes https://github.com/PistonDevelopers/texture/issues/16
